codex/requirements.txt
base58==2.1.1
PyNaCl==1.5.0
requests==2.32.3
solana==0.28.1
